Speed up alphaedges
Bunch of stuff. Most of the time is spent in fill_ov!. So rewritten the
code so that it is possible to run it in parallel. Although nothing runs
in parallel yet due to lack of @everywhere directives I think.

Also using a scratch that is filled with the alpha edge variables. It
does not speed up things by much, but makes it easier to keep the solver
hyperparameters in one place.